+++ moldesign/helpers/__init__.py
@@ -1,4 +1,4 @@
 from .helpers import *
 from .pdb import *
 from .qmmm import *
-from .widgets import *
+from .logs import *
