+++ moldesign/widgets.py
@@ -16,16 +16,56 @@ standard_library.install_aliases()
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .helpers.widgets import nbmolviz_installed
+import imp
+import functools
+from .utils import exports, exports_names
 
-if nbmolviz_installed:
+__all__ = 'BondSelector GeometryBuilder ResidueSelector Symmetrizer AtomSelector'.split()
+
+
+try:
+    import nbmolviz.widget_utils
+except ImportError:
+    nbmolviz_enabled = False
+else:
+    nbmolviz_enabled = nbmolviz.widget_utils.can_use_widgets()
+
+
+def not_installed_method(*args, **kwargs):
+    raise ImportError(
+            "ERROR: notebook visualization library not installed. "
+            "To use MDT's graphical user interfaces in a Notebook, please install "
+            "the nbmolviz library by running `pip install nbmolviz` or `conda install nbmolviz`")
+
+
+if nbmolviz_enabled:
+    # TODO: Make these into lazy imports
     from nbmolviz.widgets import (BondSelector, GeometryBuilder, ResidueSelector, Symmetrizer,
                                   AtomSelector)
     from nbmolviz.mdtconfig.compute import configure
     about = configure
 else:
-    from .helpers.widgets import not_installed_method
     BondSelector = GeometryBuilder = ResidueSelector = AtomSelector = Symmetrizer = configure \
         = about = not_installed_method
 
-__all__ = 'BondSelector GeometryBuilder ResidueSelector Symmetrizer AtomSelector'.split()
+
+def _get_nbmethod(name):
+    # don't import nbmolviz methods until a method is actually called
+    from nbmolviz import methods as nbmethods
+    module = nbmethods
+    for item in name.split('.'):
+        module = getattr(module, item)
+    return module
+
+
+@exports
+class WidgetMethod(object):
+    def __init__(self, name):
+        self.name = name
+        self.method = None
+
+    def __get__(self, instance, owner):
+        if self.method is None:
+            self.method = _get_nbmethod(self.name)
+        return functools.partial(self.method, instance)
+
